vendor: PixelLauncher: Add overlay for wallpaper app
This commit adds a 'static' overlay to remove the specification of the wallpaper
picker package that Google in all their wisdom decided to add.

The problem? We ship with WallpaperPicker2 but we weren't able to use it because
Google had their proprietary crap set.

What's a static overlay? Is an overlay that if set to true is enabled by default,
perfect for a scenario like ours.

https://github.com/DirtyUnicorns/android_frameworks_base/blob/q10x/core/java/android/content/om/OverlayInfo.java#L190-L197

Signed-off-by: smokey18 <kmboplays@gmail.com>